style: rename 'data_generation' to 'datasets'
